+++ lib/nadia/behaviour.ex
@@ -86,6 +86,46 @@ defmodule Nadia.Behaviour do
   @callback send_voice(integer, binary, [{atom, any}]) :: {:ok, Message.t()} | {:error, Error.t()}
   @callback send_voice(Client.t(), integer | binary, binary, [{atom, any}]) ::
               {:ok, Message.t()} | {:error, Error.t()}
+  @callback send_video_note(integer | binary, binary, [{atom, any}]) ::
+              {:ok, Message.t()} | {:error, Error.t()}
+  @callback send_video_note(Client.t(), integer | binary, binary, [{atom, any}]) ::
+              {:ok, Message.t()} | {:error, Error.t()}
+  @callback send_live_photo(integer | binary, binary, binary, [{atom, any}]) ::
+              {:ok, Message.t()} | {:error, Error.t()}
+  @callback send_live_photo(Client.t(), integer | binary, binary, binary, [{atom, any}]) ::
+              {:ok, Message.t()} | {:error, Error.t()}
+  @callback send_media_group(integer | binary, list | map | struct | binary, [{atom, any}]) ::
+              {:ok, [Message.t()]} | {:error, Error.t()}
+  @callback send_media_group(Client.t(), integer | binary, list | map | struct | binary, [
+              {atom, any}
+            ]) ::
+              {:ok, [Message.t()]} | {:error, Error.t()}
+  @callback send_paid_media(integer | binary, integer, list | map | struct | binary, [
+              {atom, any}
+            ]) ::
+              {:ok, Message.t()} | {:error, Error.t()}
+  @callback send_paid_media(Client.t(), integer | binary, integer, list | map | struct | binary, [
+              {atom, any}
+            ]) ::
+              {:ok, Message.t()} | {:error, Error.t()}
+  @callback send_poll(integer | binary, binary, [{atom, any}] | map) ::
+              {:ok, Message.t()} | {:error, Error.t()}
+  @callback send_poll(Client.t(), integer | binary, binary, [{atom, any}] | map) ::
+              {:ok, Message.t()} | {:error, Error.t()}
+  @callback send_dice(integer | binary) :: {:ok, Message.t()} | {:error, Error.t()}
+  @callback send_dice(integer | binary, [{atom, any}]) ::
+              {:ok, Message.t()} | {:error, Error.t()}
+  @callback send_dice(Client.t(), integer | binary) ::
+              {:ok, Message.t()} | {:error, Error.t()}
+  @callback send_dice(Client.t(), integer | binary, [{atom, any}]) ::
+              {:ok, Message.t()} | {:error, Error.t()}
+  @callback send_message_draft(integer | binary, integer) :: :ok | {:error, Error.t()}
+  @callback send_message_draft(integer | binary, integer, [{atom, any}]) ::
+              :ok | {:error, Error.t()}
+  @callback send_message_draft(Client.t(), integer | binary, integer) ::
+              :ok | {:error, Error.t()}
+  @callback send_message_draft(Client.t(), integer | binary, integer, [{atom, any}]) ::
+              :ok | {:error, Error.t()}
   @callback send_animation(integer, binary, [{atom, any}]) ::
               {:ok, Message.t()} | {:error, Error.t()}
   @callback send_animation(Client.t(), integer | binary, binary, [{atom, any}]) ::
@@ -393,6 +433,15 @@ defmodule Nadia.Behaviour do
                       send_sticker: 4,
                       send_video: 4,
                       send_voice: 4,
+                      send_video_note: 4,
+                      send_live_photo: 5,
+                      send_media_group: 4,
+                      send_paid_media: 5,
+                      send_poll: 4,
+                      send_dice: 2,
+                      send_dice: 3,
+                      send_message_draft: 3,
+                      send_message_draft: 4,
                       send_animation: 4,
                       send_location: 5,
                       send_venue: 7,